metiq: restructure of test files * moved test to separate folder * removed OLD_FRAME and BLACK_FRAME
